+++ android_repo/presentation/src/main/java/com/d101/presentation/calendar/CalendarFragment.kt
@@ -1,4 +1,4 @@
-package com.d101.presentation.main.fragments
+package com.d101.presentation.calendar
 
 import android.app.Dialog
 import android.content.Intent
