+++ android_repo/presentation/src/main/java/com/d101/presentation/leaf/LeafMessageToSendFragment.kt
@@ -12,7 +12,6 @@ import androidx.fragment.app.viewModels
 import com.d101.presentation.R
 import com.d101.presentation.databinding.FragmentLeafSendBinding
 import com.d101.presentation.main.MainActivity
-import com.d101.presentation.main.state.LeafSendViewState
 import dagger.hilt.android.AndroidEntryPoint
 import utils.CustomToast
 import utils.repeatOnStarted
