+++ android_repo/presentation/src/main/java/com/d101/presentation/tree/TreeFragmentViewState.kt
@@ -1,4 +1,4 @@
-package com.d101.presentation.main.state
+package com.d101.presentation.tree
 
 sealed class TreeFragmentViewState {
     abstract val todayDate: String
