+++ android_repo/presentation/src/main/java/com/d101/presentation/tree/TreeViewModel.kt
@@ -10,8 +10,6 @@ import com.d101.domain.model.status.TreeErrorStatus
 import com.d101.domain.usecase.main.GetMessageFromTreeUseCase
 import com.d101.domain.usecase.main.GetTodayFruitUseCase
 import com.d101.domain.usecase.usermanagement.ManageUserStatusUseCase
-import com.d101.presentation.main.state.TreeFragmentViewState
-import com.d101.presentation.main.state.TreeMessageState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
