+++ android_repo/presentation/src/main/res/layout/fragment_calendar.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".main.fragments.CalendarFragment">
+    tools:context=".calendar.CalendarFragment">
 
     <data>
